Feat: 명함방명록 삭제, 고정, 숨기기는 내가 작성한 것만 가능하도록 수정

+++ src/modules/name-card-comments/name-card-comments.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
+import { HttpErrorByCode } from '@nestjs/common/utils/http-error-by-code.util';
 import { InjectRepository } from '@nestjs/typeorm';
 import { NameCardComment } from 'src/entities/name-card-comment.entity';
 import { Repository } from 'typeorm';
@@ -36,15 +37,44 @@ export class NameCardCommentsService {
     return { nameCardCommentId: result.id };
   }
 
-  async fixNameCardComment({ id, isFix }) {
+  async fixNameCardComment(id: number, isFix: boolean, userId: number) {
+    if (!(await this.#checkCommentInMyNameCard(userId, id))) {
+      throw new ForbiddenException();
+    }
+
     await this.nameCardCommentRepository.update(id, { isFix });
   }
 
-  async privatizeNameCardComment({ id, isPrivate }) {
+  async privatizeNameCardComment(
+    id: number,
+    isPrivate: boolean,
+    userId: number,
+  ) {
+    if (!(await this.#checkCommentInMyNameCard(userId, id))) {
+      throw new ForbiddenException();
+    }
+
     await this.nameCardCommentRepository.update(id, { isPrivate });
   }
 
-  async deleteNameCardComment({ id }) {
+  async deleteNameCardComment(id: number, userId: number) {
+    if (!(await this.#checkCommentInMyNameCard(userId, id))) {
+      throw new ForbiddenException();
+    }
+
     await this.nameCardCommentRepository.delete(id);
+  }
+
+  async #checkCommentInMyNameCard(userId, nameCardCommentId): Promise<boolean> {
+    const nameCard = await this.nameCardCommentRepository.findOne({
+      where: { id: nameCardCommentId },
+      relations: ['nameCard'],
+    });
+
+    if (nameCard.userId === userId) {
+      return true;
+    } else {
+      return false;
+    }
   }
 }
